+++ app/javascript/js/organizations/services/embedded_app/components/PlanningTable.jsx
