+++ app/javascript/packs/organizations/application.js
@@ -1,3 +1,2 @@
 import '../application';
 import '../../stylessheets/organizations/organization_members.scss';
-import '../../js/organizations/planning/index'
