+++ app/javascript/packs/organizations/embedded_apps/services_overview.jsx
@@ -1,0 +1,5 @@
+import React, { h } from 'preact';
+import App from '../../../js/organizations/services/embedded_app/components/App';
+import EmbeddedApp from '../../../js/shared/EmbeddedApp';
+
+new EmbeddedApp('#planning-app', App).install();
