+++ app/views/organizations/services/index.html.slim
@@ -1,1 +1,4 @@
+- content_for(:head_additions) do
+  = javascript_pack_tag 'organizations/embedded_apps/services_overview'
+
 #planning-app
