+++ config/locales/de-CH.yml
@@ -201,6 +201,7 @@ de-CH:
         service_specifications: Pflichtenhefter
         sign_out: Abmelden
         planning: "Planung"
+        services_overview: "Einsatzübersicht"
     services:
       short_info_cell:
         service_confirmed:
