+++ config/locales/fr-CH.yml
@@ -201,6 +201,7 @@ fr-CH:
         service_specifications: cahiers des charges
         sign_out: Déconnexion
         planning: "Planification"
+        services_overview: "Aperçu d'affectations"
     services:
       short_info_cell:
         service_confirmed:
